some changes

+++ src/main/java/com/java/team/shippingservice/repository/RoleRepository.java
@@ -3,7 +3,6 @@ package com.java.team.shippingservice.repository;
 import com.java.team.shippingservice.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository<Role, Long> {
-
+public interface RoleRepository extends JpaRepository<Role, Integer> {
     Role findByCode(String code);
 }
